backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
numpy==1.26.4
scipy==1.13.0
Pillow==10.3.0
anthropic==0.28.0
python-dotenv==1.0.1
